requirements.txt
django
djangorestframework
pyyaml
requests
python-dotenv
django-cors-headers
mysqlclient